Extract AttachmentAccessLimitedUpdater logic into a worker
